First set of changes getting cint to work with size_t

+++ cint/cint/inc/DataMbr.h
@@ -53,7 +53,7 @@ G__DataMemberInfo {
   long Offset() ;
   int Bitfield();
   int ArrayDim() ;
-  long MaxIndex(int dim) ;
+  size_t MaxIndex(int dim) ;
   G__ClassInfo* MemberOf();
   void SetGlobalcomp(G__SIGNEDCHAR_T globalcomp);
   int IsValid();
