+++ cint/cint/src/DataMbr.cxx
@@ -226,7 +226,7 @@ int Cint::G__DataMemberInfo::ArrayDim()
   }
 }
 ///////////////////////////////////////////////////////////////////////////
-long Cint::G__DataMemberInfo::MaxIndex(int dim)
+size_t Cint::G__DataMemberInfo::MaxIndex(int dim)
 {
   if (IsValid()) {
     struct G__var_array* var = (struct G__var_array*) handle;
